Update Implementation.py
Im updating a few lanes to have a better understanding of the material.

Corrections & Improvements:
Clarify the __str__ Method Explanation

You wrote that __str__ prints the class attributes in string format. While this is true, it’s more precise to say that __str__ defines how an instance of MyArray is represented when converted to a string (e.g., when using print(arr)).
Suggested rewording:
The __str__ method is called when print(instance) is used. By overriding it, we customize how our object is represented as a string. Here, we return the dictionary representation of our class attributes (length and data).

Correction in the insert Method

When you insert an element, you need to shift elements from the last index to the given index, but your implementation doesn't properly maintain indexing.
The correct version should be:

def insert(self, index, item):
    if index > self.length: 
        raise IndexError("Index out of bounds")  # Prevent inserting at an invalid index
    self.length += 1
    for i in range(self.length - 1, index, -1):
        self.data[i] = self.data[i - 1]  # Shift elements to the right
    self.data[index] = item  # Insert new element at the specified index
You may also want to handle cases where index is greater than length.